+++ src/platform_impl/android/mod.rs
@@ -440,7 +440,6 @@ impl<T: 'static> EventLoop<T> {
                                             KeyAction::Up => event::ElementState::Released,
                                             _ => event::ElementState::Released,
                                         };
-                                        #[allow(deprecated)]
                                         let event = event::Event::WindowEvent {
                                             window_id,
                                             event: event::WindowEvent::KeyboardInput {
@@ -451,7 +450,6 @@ impl<T: 'static> EventLoop<T> {
                                                     virtual_keycode: ndk_keycode_to_virtualkeycode(
                                                         key.key_code(),
                                                     ),
-                                                    modifiers: event::ModifiersState::default(),
                                                 },
                                                 is_synthetic: false,
                                             },