+++ src/platform_impl/linux/wayland/seat/pointer/data.rs
@@ -9,18 +9,12 @@ use sctk::reexports::protocols::unstable::pointer_constraints::v1::client::zwp_p
 use sctk::reexports::protocols::unstable::pointer_constraints::v1::client::zwp_confined_pointer_v1::ZwpConfinedPointerV1;
 use sctk::reexports::protocols::unstable::pointer_constraints::v1::client::zwp_locked_pointer_v1::ZwpLockedPointerV1;
 
-use crate::event::{ModifiersState, TouchPhase};
+use crate::event::TouchPhase;
 
 /// A data being used by pointer handlers.
 pub(super) struct PointerData {
     /// Winit's surface the pointer is currently over.
     pub surface: Option<WlSurface>,
-
-    /// Current modifiers state.
-    ///
-    /// This refers a state of modifiers from `WlKeyboard` on
-    /// the given seat.
-    pub modifiers_state: Rc<RefCell<ModifiersState>>,
 
     /// Pointer constraints.
     pub pointer_constraints: Option<Attached<ZwpPointerConstraintsV1>>,
@@ -43,7 +37,6 @@ impl PointerData {
         confined_pointer: Rc<RefCell<Option<ZwpConfinedPointerV1>>>,
         locked_pointer: Rc<RefCell<Option<ZwpLockedPointerV1>>>,
         pointer_constraints: Option<Attached<ZwpPointerConstraintsV1>>,
-        modifiers_state: Rc<RefCell<ModifiersState>>,
     ) -> Self {
         Self {
             surface: None,
@@ -51,7 +44,6 @@ impl PointerData {
             latest_enter_serial: Rc::new(Cell::new(0)),
             confined_pointer,
             locked_pointer,
-            modifiers_state,
             pointer_constraints,
             axis_data: AxisData::new(),
         }