+++ src/platform_impl/macos/event.rs
@@ -287,14 +287,12 @@ pub unsafe fn modifier_event(
 
         let scancode = get_scancode(ns_event);
         let virtual_keycode = scancode_to_keycode(scancode);
-        #[allow(deprecated)]
         Some(WindowEvent::KeyboardInput {
             device_id: DEVICE_ID,
             input: KeyboardInput {
                 state,
                 scancode: scancode as _,
                 virtual_keycode,
-                modifiers: event_mods(ns_event),
             },
             is_synthetic: false,
         })