+++ src/platform_impl/web/web_sys/canvas.rs
@@ -3,7 +3,7 @@ use super::event_handle::EventListenerHandle;
 use super::media_query_handle::MediaQueryListHandle;
 use crate::dpi::{LogicalPosition, PhysicalPosition, PhysicalSize};
 use crate::error::OsError as RootOE;
-use crate::event::{ModifiersState, MouseButton, MouseScrollDelta, ScanCode, VirtualKeyCode};
+use crate::event::{MouseButton, MouseScrollDelta, ScanCode, VirtualKeyCode};
 use crate::platform_impl::{OsError, PlatformSpecificWindowBuilderAttributes};
 
 use std::cell::RefCell;
@@ -150,24 +150,20 @@ impl Canvas {
 
     pub fn on_keyboard_release<F>(&mut self, mut handler: F)
     where
-        F: 'static + FnMut(ScanCode, Option<VirtualKeyCode>, ModifiersState),
+        F: 'static + FnMut(ScanCode, Option<VirtualKeyCode>),
     {
         self.on_keyboard_release = Some(self.common.add_user_event(
             "keyup",
             move |event: KeyboardEvent| {
                 event.prevent_default();
-                handler(
-                    event::scan_code(&event),
-                    event::virtual_key_code(&event),
-                    event::keyboard_modifiers(&event),
-                );
+                handler(event::scan_code(&event), event::virtual_key_code(&event));
             },
         ));
     }
 
     pub fn on_keyboard_press<F>(&mut self, mut handler: F)
     where
-        F: 'static + FnMut(ScanCode, Option<VirtualKeyCode>, ModifiersState),
+        F: 'static + FnMut(ScanCode, Option<VirtualKeyCode>),
     {
         self.on_keyboard_press = Some(self.common.add_user_event(
             "keydown",
@@ -183,11 +179,7 @@ impl Canvas {
                 if !is_key_string || is_shortcut_modifiers {
                     event.prevent_default();
                 }
-                handler(
-                    event::scan_code(&event),
-                    event::virtual_key_code(&event),
-                    event::keyboard_modifiers(&event),
-                );
+                handler(event::scan_code(&event), event::virtual_key_code(&event));
             },
         ));
     }
@@ -233,7 +225,7 @@ impl Canvas {
 
     pub fn on_mouse_release<F>(&mut self, handler: F)
     where
-        F: 'static + FnMut(i32, MouseButton, ModifiersState),
+        F: 'static + FnMut(i32, MouseButton),
     {
         match &mut self.mouse_state {
             MouseState::HasPointerEvent(h) => h.on_mouse_release(&self.common, handler),
@@ -243,7 +235,7 @@ impl Canvas {
 
     pub fn on_mouse_press<F>(&mut self, handler: F)
     where
-        F: 'static + FnMut(i32, PhysicalPosition<f64>, MouseButton, ModifiersState),
+        F: 'static + FnMut(i32, PhysicalPosition<f64>, MouseButton),
     {
         match &mut self.mouse_state {
             MouseState::HasPointerEvent(h) => h.on_mouse_press(&self.common, handler),
@@ -253,7 +245,7 @@ impl Canvas {
 
     pub fn on_cursor_move<F>(&mut self, handler: F)
     where
-        F: 'static + FnMut(i32, PhysicalPosition<f64>, PhysicalPosition<f64>, ModifiersState),
+        F: 'static + FnMut(i32, PhysicalPosition<f64>, PhysicalPosition<f64>),
     {
         match &mut self.mouse_state {
             MouseState::HasPointerEvent(h) => h.on_cursor_move(&self.common, handler),
@@ -263,12 +255,12 @@ impl Canvas {
 
     pub fn on_mouse_wheel<F>(&mut self, mut handler: F)
     where
-        F: 'static + FnMut(i32, MouseScrollDelta, ModifiersState),
+        F: 'static + FnMut(i32, MouseScrollDelta),
     {
         self.on_mouse_wheel = Some(self.common.add_event("wheel", move |event: WheelEvent| {
             event.prevent_default();
             if let Some(delta) = event::mouse_scroll_delta(&event) {
-                handler(0, delta, event::mouse_modifiers(&event));
+                handler(0, delta);
             }
         }));
     }